refactor estimate functions into analysis/stitch.py
